requirements.txt
Flask
gunicorn
python-dotenv
sqlalchemy
pymysql
python-dotenv
